requirements.txt
pymongo[srv]==3.6
dnspython==1.16.0
python-dotenv
pymongo
certifi
xgboost
PyYAML
scikit
numpy
pandas
mlflow
dagshub